requirements.txt
pycurl
certifi
discord-interactions
datetime
pytz